StackFusion/requirements.txt
django==3.0.4
anyjson==0.3.3
asgiref==3.2.3
dj-config-url==0.1.1
dj-database-url==0.5.0
django-appconf==1.0.3
django-celery==3.3.1
django-celery-email
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.5.0
pip==19.0.3
psycopg2-binary==2.8.4
pytz==2019.3
setuptools==40.8.0
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
